pyproject.toml
[project]
name = "hansard-spotlight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.2",
    "fire>=0.7.0",
    "ibis-framework[duckdb,examples]>=10.6.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "marimo[sql]>=0.13.10",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "opencv-python>=4.11.0.86",
    "pdfquery>=0.4.3",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "pypdf>=5.5.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "rake-nltk>=1.0.6",
    "rich>=14.0.0",
    "six>=1.17.0",
    "spacy>=3.8.4",
    "tabulate>=0.9.0",
    "toolz>=1.0.0",
]

[dependency-groups]
dev = [
    "icecream>=2.1.4",
    "ipython>=9.2.0",
    "mypy>=1.16.1",
    "pymupdf-stubs>=1.26.1.post1",
    "pysnooper>=1.2.3",
    "rich>=14.0.0",
    "wat-inspector>=0.4.3",
]
